Platform/Sgi: non-discoverable IO virt block on rdn2 platform
The RD-N2 platform uses one instance of the IO Virtualization block
to connect PL011 UART controllers and PL330 DMA controllers. Describe
these devices by including the SsdtIoVirtBlk.asl ACPI table. In addition
to this, add named-component IORT nodes for DMA PL330 DMA controllers
that is interfaced with the SMMUv3 present in the IO virtualization
block that is used to connect non-discoverable devices. This node maps
to the DMA PL330 device node present in Ssdt table for the platform.

Signed-off-by: Vivek Gautam <vivek.gautam@arm.com>
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I2a9c73e949524c9f15bd18fca251101929181376