Platform/Sgi: Add rdn2cfg1 platform specific acpi tables
The IO topology on the RD-N2-Cfg1 platform is built using the IO
Virtualization block. There are two instances of the IO Virtualization
block on the RD-N2 platform and one of these instances is used to
connect to a PCIe root bridge. Add RD-N2-Cfg1 platform specific IORT,
MCFG and SSDT ACPI tables to represent this IO topology.

Signed-off-by: Vivek Gautam <vivek.gautam@arm.com>
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I73821600d58811c641f2bb5c2da0a492c94ee251